Disabled multishard cmds and test for HTTP/WS (#1363)
