+++ internal/server/httpws/redisCmdAdapter.go
@@ -1,16 +1,17 @@
-package utils
+package httpws
 
 import (
 	"encoding/base64"
 	"encoding/json"
 	"errors"
 	"fmt"
-
 	"io"
 	"net/http"
 	"regexp"
 	"strconv"
 	"strings"
+
+	"github.com/dicedb/dice/internal/server/utils"
 
 	"github.com/dicedb/dice/internal/cmd"
 	diceerrors "github.com/dicedb/dice/internal/errors"
@@ -64,7 +65,7 @@ func ParseHTTPRequest(r *http.Request) (*cmd.DiceDBCmd, error) {
 	queryParams := r.URL.Query()
 	keyPrefix := queryParams.Get(KeyPrefix)
 
-	if keyPrefix != "" && command == JSONIngest {
+	if keyPrefix != "" && command == utils.JSONIngest {
 		args = append(args, keyPrefix)
 	}
 	// Step 1: Handle JSON body if present
@@ -173,7 +174,7 @@ func ParseWebsocketMessage(msg []byte) (*cmd.DiceDBCmd, error) {
 
 	// if key prefix is empty for JSON.INGEST command
 	// add "" to cmdArr
-	if command == JSONIngest && len(cmdArr) == 2 {
+	if command == utils.JSONIngest && len(cmdArr) == 2 {
 		cmdArr = append([]string{""}, cmdArr...)
 	}
